Add files via upload
Has to come with a data folder in order to work, but that one is currently a bit too big to have in github. Also has no direct input on what data to use, so it relies on hardcoding for that, which makes it unpractical unless you have a proper visual studios 2019 project set up already.